+++ marching_cube.h
@@ -13,16 +13,16 @@ class MarchingCube {
 	//Code From by http://paulbourke.net/geometry/polygonise/
 	public:
 		ThreeDVector* min_corner;
-		long double size;
+		float size;
 
-		MarchingCube(ThreeDVector* min_corner, long double size);
+		MarchingCube(ThreeDVector* min_corner, float size);
 		~MarchingCube();
-		vector<vector<pair<ThreeDVector*, ThreeDVector*> > >* triangulate(ParticleGrid* particle_grid, long double isovalue_threshold);
-		pair<ThreeDVector*, ThreeDVector*> interpolatePoint(int p1_index, int p2_index, long double color_p1, long double color_p2, long double isovalue_threshold, ParticleGrid* particle_grid);
+		vector<vector<pair<ThreeDVector*, ThreeDVector*> > >* triangulate(ParticleGrid* particle_grid, float isovalue_threshold);
+		pair<ThreeDVector*, ThreeDVector*> interpolatePoint(int p1_index, int p2_index, float color_p1, float color_p2, float isovalue_threshold, ParticleGrid* particle_grid);
 		ThreeDVector* pointAt(int corner_index);
 
-		static vector<MarchingCube*>* generateGrid(vector<Particle*>* particles, long double step_size);
-		//static vector<MarchingCube*>* generateGridFast(ParticleGrid* particle_grid, long double step_size);
+		static vector<MarchingCube*>* generateGrid(vector<Particle*>* particles, float step_size);
+		//static vector<MarchingCube*>* generateGridFast(ParticleGrid* particle_grid, float step_size);
 
 		static const int edgeTable[256];
 		static const int triTable[256][16];
